finished

+++ server.py
@@ -2,14 +2,13 @@
 
 import argparse
 import mimetypes
-from pathlib import Path
 import socket
 import threading
 
 from http.enums import HttpMethod
 from http.header import HttpHeader, HEADER_CONTENT_TYPE
 from http.request import HttpRequest
-from http.response import HttpResponse, HttpResponseError, HttpResponseNotFound, HttpResponseUnsupportedMediaType
+from http.response import HttpResponse, HttpResponseError, HttpResponseMethodNotAllowed, HttpResponseNotFound, HttpResponseUnsupportedMediaType
 from settings import DEFAULT_PORT, VHOSTS_FILE
 from utils.entity import generate_output
 from utils.vhosts import Vhost
@@ -45,24 +44,17 @@ class Server:
         self.__socket.close()
         self.__socket = None
 
-
-    @staticmethod
-    def get_file_path(request):
-        file_path = request.get_vhost().get_host_root_path().joinpath(request.get_path())
-        if not file_path.is_file():
-            file_path = file_path.joinpath(request.get_vhost().get_index_file())
-
-        if not file_path.exists():
-            raise HttpResponseNotFound(content="File not found")
-        return file_path
-
-
     @staticmethod
     def __get_response(request: HttpRequest) -> HttpResponse:
         response = HttpResponse()
 
         if request.get_method() == HttpMethod.GET:
-            file_path = Server.get_file_path(request)
+            file_path = request.get_vhost().get_host_root_path().joinpath(request.get_path())
+            if not file_path.is_file():
+                file_path = file_path.joinpath(request.get_vhost().get_index_file())
+
+            if not file_path.exists():
+                raise HttpResponseNotFound(content="File not found")
 
             content = Vhost.get_file_contents(file_path)
             response = HttpResponse(content=content)
@@ -78,12 +70,15 @@ class Server:
             # TODO
             pass
         elif request.get_method() == HttpMethod.DELETE:
-            # Also checks for file existence
-            file_path = Server.get_file_path(request)
+            file_path = request.get_vhost().get_host_root_path().joinpath(request.get_path())
+            if not file_path.exists():
+                raise HttpResponseNotFound(content="File not found")
+
+            if not file_path.is_file():
+                raise HttpResponseMethodNotAllowed()
             # Deletes files and also a folder if it's empty
             Vhost.delete_file(file_path, request.get_vhost().get_host_root_path())
 
-            
         elif request.get_method() == HttpMethod.NTW22INFO:
             # TODO
             pass
